Complete project cleanup, faster changed fields, fixed changed fields for JSON field
