+++ .config/flake8
@@ -1,0 +1,2 @@
+[flake8]
+max-line-length = 120
