+++ chamber/storages/boto3.py
@@ -1,7 +1,6 @@
 import boto3
 from botocore.client import Config
 
-from django.conf import settings
 from django.core.files.base import ContentFile
 from django.core.files.storage import DefaultStorage
