+++ example/dj/apps/test_chamber/importers.py
@@ -1,3 +1,7 @@
+import os
+
+from django.conf import settings
+
 from chamber.importers import BulkCSVImporter, CSVImporter
 
 from .models import CSVRecord
@@ -6,7 +10,7 @@ from .models import CSVRecord
 class BulkCSVRecordImporter(BulkCSVImporter):
     model_class = CSVRecord
     fields = ('id', 'name', 'number')
-    csv_path = 'data/all_fields_filled.csv'
+    csv_path = os.path.join(settings.PROJECT_DIR, 'data', 'all_fields_filled.csv')
 
     def clean_number(self, value):
         # Just to test clean methods are called
@@ -16,7 +20,7 @@ class BulkCSVRecordImporter(BulkCSVImporter):
 class CSVRecordImporter(CSVImporter):
     model_class = CSVRecord
     fields = ('id', 'name', 'number')
-    csv_path = 'data/all_fields_filled.csv'
+    csv_path = os.path.join(settings.PROJECT_DIR, 'data', 'all_fields_filled.csv')
 
     def clean_number(self, value):
         # Just to test clean methods are called
