+++ example/dj/apps/test_chamber/management/commands/bulk_csv_import.py
@@ -1,3 +1,7 @@
+import os
+
+from django.conf import settings
+
 from chamber.commands import BulkImportCSVCommand
 
 from test_chamber.models import CSVRecord  # pylint: disable=E0401
@@ -6,7 +10,7 @@ from test_chamber.models import CSVRecord  # pylint: disable=E0401
 class Command(BulkImportCSVCommand):
     model_class = CSVRecord
     fields = ('id', 'name', 'number')
-    csv_path = 'data/all_fields_filled.csv'
+    csv_path = os.path.join(settings.PROJECT_DIR, 'data', 'all_fields_filled.csv')
 
     def clean_number(self, value):
         # Just to test clean methods are called
