+++ example/requirements.txt
@@ -1,9 +1,9 @@
 Django==3.1.13
+flake8
+coveralls
 diff-match-patch==20110725.1
 django-germanium==2.1.0
 six==1.10.0
-coverage==5.3.1
-coveralls==2.2.0
 Pillow==8.3.2
 boto3==1.16.47
 django-storages==1.11.1
